numpy's linear algebra stuff no longer casts object numpy arrays to float, so I have to do it. Dumb.

+++ skpp/skpp.py
@@ -307,7 +307,8 @@ class ProjectionPursuitRegressor(BaseEstimator, TransformerMixin, RegressorMixin
 				b = -sum([self._out_dim_weights[k] * self._beta[k, j] *
 					numpy.dot(J.T, G_j[:, k]) for k in range(Y.shape[1])])
 
-				delta = numpy.linalg.lstsq(A, b, rcond=-1)[0]
+				delta = numpy.linalg.lstsq(A.astype(numpy.float64, copy=False),
+					b.astype(numpy.float64, copy=False), rcond=-1)[0]
 				# TODO implement halving step if the loss doesn't decrease with
 				# this update.
 				alpha = self._alpha[:, j] + delta
@@ -385,7 +386,9 @@ class ProjectionPursuitRegressor(BaseEstimator, TransformerMixin, RegressorMixin
 
 		"""
 		if self.fit_type == 'polyfit':
-			coeffs = numpy.polyfit(x, y, deg=self.degree, w=self._example_weights)
+			coeffs = numpy.polyfit(x.astype(numpy.float64, copy=False),
+				y.astype(numpy.float64, copy=False), deg=self.degree,
+				w=self._example_weights)
 			fit = numpy.poly1d(coeffs)
 			deriv = fit.deriv(m=1)
 		elif self.fit_type == 'spline':
